refactor: staleTime 상수로 관리
