+++ packages/mobile/src/consts/react-query/queryKey.ts
