+++ packages/mobile/src/consts/react-query/staleTime.ts
@@ -1,0 +1,5 @@
+export const staleTime = {
+  short: 1000 * 5,
+  basic: 1000 * 30,
+  long: 1000 * 300,
+};
