+++ packages/mobile/src/hooks/api/board/index.ts
@@ -4,7 +4,8 @@ import {
   useCoreQuery,
 } from "@hooks/api/core";
 import { BoardApiService } from "@shared/swagger-api/generated";
-import { queryKey } from "src/consts/react-query";
+import { queryKey } from "src/consts/react-query/queryKey";
+import { staleTime } from "src/consts/react-query/staleTime";
 import { queryClient } from "src/main";
 import { GetParams } from "src/type/utils";
 import { getUuid } from "src/utils/storage";
@@ -58,7 +59,7 @@ export const useBoardsQuery = () => {
     () => {
       return BoardApiService.boardControllerFind({ uuid });
     },
-    { staleTime: 300000 },
+    { staleTime: staleTime.long },
   );
 };
