+++ packages/mobile/src/hooks/api/cafeteria/index.ts
@@ -4,7 +4,7 @@ import {
 } from "@shared/swagger-api/generated";
 import { useQueries } from "@tanstack/react-query";
 import { CAFETERIA_MENUS } from "src/constants";
-import { queryKey } from "src/consts/react-query";
+import { queryKey } from "src/consts/react-query/queryKey";
 
 import { useCoreQuery } from "../core";
