+++ packages/mobile/src/hooks/api/school/index.ts
@@ -1,6 +1,6 @@
 import { useCoreQuery } from "@hooks/api/core";
 import { PlaceApiService } from "@shared/swagger-api/generated";
-import { queryKey } from "src/consts/react-query";
+import { queryKey } from "src/consts/react-query/queryKey";
 import { GetParams } from "src/type/utils";
 
 export const useSchoolQuery = (params?: GetParams<typeof PlaceApiService.placeControllerFindSchool>) => {
